+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 + [x] Added new languages for `i18n` component: `de`, `es`, `fr`, `hi`, `it`, `pl`, `pt`, `zh`
 + [x] Ukrainian language for `i18n` component renamed to `uk`
 + [x] Updated `hotkey` component: arbitrary order of special keys: `ctrl+shift+alt+f`, added handler for `windows`/`command` key
++ [x] Fixed icons for `wizard`
++ [x] Added component `wizard-classic`
 
 ### 5.1.1
 + [x] Added a combined hotkeys feature. Now you can create hotkey as: `alt+1 2`